fix (createNonce) validate sig
